+++ src/app/libs/gateway.js
@@ -5,10 +5,9 @@
 import str from 'string';
 const debug = require('debug')('AiC:Libs:Gateway');
 
-// APP
-import RestAPI from 'app/libs/rest-api';
-
 const Gateway = {
+
+	requestAPI: {},
 
 	register(options) {
 		debug('register', options);
@@ -16,6 +15,9 @@ const Gateway = {
 		options.actions.forEach(action => {
 			debug('actions', action);
 			action.namespace = options.namespace;
+			if (Gateway[options.namespace].hasOwnProperty(action.action.name)) {
+				throw new Error('Duplicated action on Gateway namespace. ', action.action.name, 'already exists in', options.namespace);
+			}
 			Gateway[options.namespace][action.action.name] = Gateway.request.bind(Gateway, action);
 		});
 	},
@@ -25,15 +27,37 @@ const Gateway = {
 		Gateway.adapters = adapters;
 	},
 
+	getAdapter(namespace, action, type) {
+		if (Gateway.adapters[namespace] &&
+				Gateway.adapters[namespace][action] &&
+				Gateway.adapters[namespace][action][type]) {
+			return Gateway.adapters[namespace][action][type];
+		}
+		return false;
+	},
+
 	request(options, obj) {
 		debug(options, obj);
-		const requestAdapter = Gateway.adapters[options.namespace][options.action.name] ? Gateway.adapters[options.namespace][options.action.name].request : false;
-		const responseAdapter = Gateway.adapters[options.namespace][options.action.name] ? Gateway.adapters[options.namespace][options.action.name].response : false;
-		const schemaAdapter = Gateway.adapters[options.namespace][options.action.name] ? Gateway.adapters[options.namespace][options.action.name].schema : false;
+		if (options.placeholder) {
+			return options.placeholder();
+		}
+		const requestAdapter = this.getAdapter(options.namespace, options.action.name, 'request');
+		const responseAdapter = this.getAdapter(options.namespace, options.action.name, 'response');
+		const schemaAdapter = this.getAdapter(options.namespace, options.action.name, 'schema');
 		const optionsAPI = {
-			pathname: str(options.pathname).template(obj).s,
 			method: options.action.method
 		};
+		const requestCallerName = options.requestAPI ? options.requestAPI : 'default';
+		if (typeof Gateway.requestAPI[requestCallerName] !== 'function') {
+			throw new Error('Missing or not a function Gateway.requestAPI[', requestCallerName, ']');
+		}
+		const requestCaller = Gateway.requestAPI[requestCallerName];
+		if (options.pathname) {
+			optionsAPI.pathname = str(options.pathname).template(obj).s;
+		}
+		if (options.url) {
+			optionsAPI.url = options.url;
+		}
 		if (obj && obj.file) {
 			optionsAPI.rawData = new FormData();
 			optionsAPI.rawData.append('file', obj.file);
@@ -47,9 +71,9 @@ const Gateway = {
 		}
 		debug(optionsAPI);
 		if (responseAdapter) {
-			return RestAPI.apiCallAuth(optionsAPI).then(responseAdapter);
+			return requestCaller(optionsAPI).then(responseAdapter);
 		}
-		return RestAPI.apiCallAuth(optionsAPI);
+		return requestCaller(optionsAPI);
 	}
 
 };